requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
tradingview-screener>=3.0.0
tradingview-ta>=3.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
